feat(compiler): carry integer literals as exact i128 through the pipeline
The parser parsed every integer literal into i64, so u64::MAX was an
invalid literal and i64::MIN needed a decimal-only parser fold. Literals
now carry their exact mathematical value from the AST through the checker,
HIR and SIR; physical MIR derives the destination-width bit pattern as
PhysicalConst::IntegerBits and codegen emits it unsigned.

+++ hew-types/src/check/expressions.rs
@@ -1103,7 +1103,7 @@ impl Checker {
                 continue;
             }
             if let ConstValue::Integer(v) = value {
-                env.insert(name.clone(), i128::from(*v));
+                env.insert(name.clone(), *v);
             }
         }
         env
